+++ src/lib.rs
@@ -1,4 +1,5 @@
 mod buffer_size;
+pub mod callback;
 pub mod converters;
 pub mod err;
 mod mixer;
@@ -17,30 +18,36 @@ pub use self::{
 mod tests {
     /*use std::{fs::File, io::stdin, time::Duration};
 
-    use crate::{err::Error, source::Symph, Sink, BufferSize};
+    use crate::{err::Error, source::Symph, BufferSize, Sink};
 
     use anyhow::Result;
     use cpal::traits::DeviceTrait;
 
     #[test]
     fn play_audio() -> Result<()> {
+        let home_path = include_str!("../tmp/home").to_owned();
+
         let mut sink = Sink::default();
-        let src = Symph::try_new(File::open(
-            "/mnt/x/Music/AJR - Neotheater - 01 Next Up Forever.flac",
-        )?, &Default::default())?;
+        let src = Symph::try_new(
+            File::open(
+                home_path + "/music/4tet - 4th -03 Air.mp3",
+                //HOME_PATH + "/music/AJR - Neotheater - 01 Next Up Forever.flac",
+            )?,
+            &Default::default(),
+        )?;
         sink.on_callback(Some(|c| println!("callback: {c:?}")))?;
         sink.on_err_callback(Some(|e: Error| println!("{}", e)))?;
-        sink.volume(0.2)?;
-        for i in Sink::list_devices()? {
+        sink.volume(1.)?;
+        /*for i in Sink::list_devices()? {
             println!("{}", i.name()?);
-        }
+        }*/
         sink.load(src, true)?;
         sink.set_fade_len(Duration::from_millis(200))?;
         //thread::sleep(Duration::from_secs(5));
         loop {
             let mut s = String::new();
             _ = stdin().read_line(&mut s);
-            sink.play(!sink.is_playing()?)?;
+            //sink.play(!sink.is_playing()?)?;
             let ts = sink.get_timestamp()?;
             println!("{:?}/{:?}", ts.current, ts.total);
         }